+++ src/enum_type.rs
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
 use indexmap::IndexMap;
 use regex::Regex;
